requirements.txt
flask
bcrypt
python-dotenv
requests
black
mysql-connector-python
